requirements.txt
pymysql==1.1.2
datasets==4.8.4
pandas>=2.3.0
numpy>= 2.0.0
sentence-transformers==5.4.1
python-dotenv>=1.2.0
langchain-community >= 0.4
langchain-huggingface >=1.2.0
langchain>=1.2.2
faiss-cpu>=1.12.0 # CPU 사용 시
#faiss-gpu-cu12>=1.12.0 # GPU 사용 시